requirements.txt
pygame-ce
esper == 2.5
pygbag
pyinstaller
